Shifted common tests into new test_compute module
