+++ tests/test_compute_ComputationMethod.py
@@ -1,18 +1,6 @@
 import pytest
 
-from RIAssigner.compute import ComputationMethod, CubicSpline, Kovats
-
-
-@pytest.mark.parametrize('this, other, expected', [
-    [CubicSpline(), CubicSpline(), True],
-    [Kovats(), Kovats(), True],
-    [CubicSpline(), Kovats(), False],
-    [Kovats(), object(), False],
-    [CubicSpline(), object(), False],
-])
-def test_equal(this, other, expected):
-    actual = (this == other)
-    assert actual == expected
+from RIAssigner.compute import ComputationMethod
 
 
 def test_abc():
