+++ tests/test_compute_CubicSpline.py
@@ -2,13 +2,8 @@ import numpy
 import pytest
 from RIAssigner.compute import CubicSpline
 
-from tests.fixtures import indexed_data, non_indexed_data, reference_alkanes, queries
+from tests.fixtures import reference_alkanes, queries
 from tests.fixtures.mocks import DataStub
-
-
-def test_construct():
-    method = CubicSpline()
-    assert method is not None
 
 
 @pytest.mark.parametrize('reference_points, query_points, expected', [
@@ -23,26 +18,6 @@ def test_simple_computations(reference_points, query_points, expected):
 
     actual = method.compute(query, reference)
     numpy.testing.assert_array_almost_equal(actual, expected, 4)
-
-
-def test_exception_reference_none(non_indexed_data):
-    method = CubicSpline()
-    with pytest.raises(AssertionError) as exception:
-        method.compute(non_indexed_data, None)
-
-    message = exception.value.args[0]
-    assert exception.typename == "AssertionError"
-    assert message == "Reference data is 'None'."
-
-
-def test_exception_query_none(indexed_data):
-    method = CubicSpline()
-    with pytest.raises(AssertionError) as exception:
-        method.compute(None, indexed_data)
-
-    message = exception.value.args[0]
-    assert exception.typename == "AssertionError"
-    assert message == "Query data is 'None'."
 
 
 @pytest.mark.method('cubicspline')
